refactor sys-core utilities
