+++ sys-account/service/flu/lib/view/widgets/view_model/verify_view_model.dart
@@ -67,7 +67,7 @@ class VerifyAccountViewModel extends BaseModel {
     _loadingProcess(true);
     String accountId = _accountId;
     if (accountId == null) {
-      accountId = await getTempAccountId();
+      accountId = getTempAccountId();
     }
     await authRepo.AuthRepo.verifyAccount(
             id: accountId, verificationToken: _userVerifyToken)
